refactor(storage): extract the backstage materialize client into @yucp/shared
