+++ apps/api/src/routes/packages.ts
@@ -4,18 +4,19 @@ import {
   getProviderDescriptor,
 } from '@yucp/providers/providerMetadata';
 import {
+  BackstageMaterializeError,
+  isFetchAbortOrTimeoutError,
   mergeYucpAliasPackageMetadata,
   type PublicApiScope,
+  runBackstageMaterialize,
   type YucpAliasPackageContract,
 } from '@yucp/shared';
 import type { ApiActorBinding } from '@yucp/shared/apiActor';
 import {
   type BackstageMaterializeClaims,
-  type BackstageMaterializeResult,
   type BackstageUploadClaims,
   type BackstageUploadResult,
   parseMaterializeClaims,
-  parseMaterializeResult,
   parseUploadClaims,
   parseUploadResult,
   sign,
@@ -58,9 +59,6 @@ const BACKSTAGE_REPO_TOKEN_TTL_MS = 30 * 24 * 60 * 60 * 1000;
 const DEFAULT_BACKSTAGE_LIVE_SYNC_TIMEOUT_MS = 1_500;
 const MAX_BACKSTAGE_UPLOAD_AUTHORIZATION_BYTES = 16 * 1024;
 const MAX_BACKSTAGE_PACKAGE_MEDIA_BYTES = 5 * 1024 * 1024;
-const BACKSTAGE_MATERIALIZE_REQUEST_TIMEOUT_MS = 5_000;
-const BACKSTAGE_MATERIALIZE_POLL_INTERVAL_MS = 500;
-const BACKSTAGE_MATERIALIZE_DEADLINE_MS = 20_000;
 const BACKSTAGE_PACKAGE_MEDIA_CONTENT_TYPES = new Set([
   'image/gif',
   'image/jpeg',
@@ -210,13 +208,6 @@ class BackstageUploadAuthorizationLimitError extends Error {
   }
 }
 
-class BackstageMaterializeError extends Error {
-  constructor(message: string) {
-    super(message);
-    this.name = 'BackstageMaterializeError';
-  }
-}
-
 function jsonResponse(body: object, status = 200): Response {
   return new Response(JSON.stringify(body), {
     status,
@@ -235,18 +226,6 @@ function isLoreBackstageDependencyError(error: unknown): boolean {
   );
 }
 
-function isFetchAbortOrTimeoutError(error: unknown): boolean {
-  if (!(error instanceof Error)) {
-    return false;
-  }
-  const name = error.name.trim().toLowerCase();
-  if (name === 'aborterror' || name === 'timeouterror') {
-    return true;
-  }
-  const message = error.message.trim().toLowerCase();
-  return message.includes('aborted') || message.includes('timed out');
-}
-
 function loreBackstageUnavailableResponse(error: string): Response {
   return jsonResponse({ error }, 502);
 }
@@ -257,132 +236,6 @@ export function trimTrailingForwardSlashes(value: string): string {
     end -= 1;
   }
   return end === value.length ? value : value.slice(0, end);
-}
-
-function materializeRequestSignal(deadline: number): AbortSignal {
-  return AbortSignal.timeout(
-    Math.max(1, Math.min(BACKSTAGE_MATERIALIZE_REQUEST_TIMEOUT_MS, deadline - Date.now()))
-  );
-}
-
-async function parseMaterializeJsonResponse(
-  response: Response,
-  operation: string
-): Promise<unknown> {
-  if (!response.ok) {
-    throw new BackstageMaterializeError(
-      `${operation} failed (${response.status} ${response.statusText})`
-    );
-  }
-  try {
-    return await response.json();
-  } catch {
-    throw new BackstageMaterializeError(`${operation} returned invalid JSON`);
-  }
-}
-
-async function runBackstageMaterialize(
-  config: PackagesConfig,
-  claims: BackstageMaterializeClaims
-): Promise<BackstageMaterializeResult> {
-  if (!config.backstageIngestSecret || !config.ingestBaseUrl) {
-    throw new BackstageMaterializeError('Backstage ingest service is not configured');
-  }
-  const ingestBaseUrl = trimTrailingForwardSlashes(
-    assertSecureLoreUrl(config.ingestBaseUrl, 'LORE_INGEST_BASE_URL').toString()
-  );
-  const materializeToken = await sign(config.backstageIngestSecret, claims);
-  const deadline = Date.now() + BACKSTAGE_MATERIALIZE_DEADLINE_MS;
-
-  let enqueueResponse: Response;
-  try {
-    enqueueResponse = await fetch(`${ingestBaseUrl}/materialize`, {
-      method: 'POST',
-      headers: { Authorization: `Bearer ${materializeToken}` },
-      signal: materializeRequestSignal(deadline),
-    });
-  } catch (error) {
-    throw new BackstageMaterializeError(
-      `Backstage materialize request failed: ${
-        isFetchAbortOrTimeoutError(error) ? 'request timed out' : 'sidecar unavailable'
-      }`
-    );
-  }
-  const enqueuePayload = await parseMaterializeJsonResponse(
-    enqueueResponse,
-    'Backstage materialize request'
-  );
-  if (!isRecord(enqueuePayload) || typeof enqueuePayload.jobId !== 'string') {
-    throw new BackstageMaterializeError(
-      'Backstage materialize request did not return a valid jobId'
-    );
-  }
-  const jobId = enqueuePayload.jobId.trim();
-  if (!jobId) {
-    throw new BackstageMaterializeError(
-      'Backstage materialize request did not return a valid jobId'
-    );
-  }
-  const jobUrl = `${ingestBaseUrl}/jobs/${encodeURIComponent(jobId)}`;
-
-  // ponytail: a unitypackage shim materializes in seconds with a few polls. A very large .zip
-  // repack could exceed the Worker's subrequest or time budget and would need a client-async
-  // publish handoff, which is intentionally out of scope here.
-  for (;;) {
-    if (Date.now() >= deadline) {
-      throw new BackstageMaterializeError('Timed out waiting for Backstage materialization');
-    }
-
-    let jobResponse: Response;
-    try {
-      jobResponse = await fetch(jobUrl, {
-        method: 'GET',
-        headers: { Authorization: `Bearer ${materializeToken}` },
-        signal: materializeRequestSignal(deadline),
-      });
-    } catch (error) {
-      throw new BackstageMaterializeError(
-        `Backstage materialize job polling failed: ${
-          isFetchAbortOrTimeoutError(error) ? 'request timed out' : 'sidecar unavailable'
-        }`
-      );
-    }
-    const jobPayload = await parseMaterializeJsonResponse(
-      jobResponse,
-      'Backstage materialize job polling'
-    );
-    if (!isRecord(jobPayload)) {
-      throw new BackstageMaterializeError('Backstage materialize job returned an invalid state');
-    }
-    if (jobPayload.state === 'completed') {
-      if (typeof jobPayload.result !== 'string' || !jobPayload.result.trim()) {
-        throw new BackstageMaterializeError(
-          'Completed Backstage materialize job did not return a signed result'
-        );
-      }
-      try {
-        return parseMaterializeResult(
-          await verify(config.backstageIngestSecret, jobPayload.result)
-        );
-      } catch {
-        throw new BackstageMaterializeError(
-          'Completed Backstage materialize job returned an invalid signed result'
-        );
-      }
-    }
-    if (jobPayload.state === 'failed') {
-      const reason =
-        typeof jobPayload.reason === 'string' && jobPayload.reason.trim()
-          ? jobPayload.reason.trim()
-          : 'unknown failure';
-      throw new BackstageMaterializeError(`Backstage materialize job failed: ${reason}`);
-    }
-    if (jobPayload.state !== 'processing') {
-      throw new BackstageMaterializeError('Backstage materialize job returned an invalid state');
-    }
-
-    await new Promise((resolve) => setTimeout(resolve, BACKSTAGE_MATERIALIZE_POLL_INTERVAL_MS));
-  }
 }
 
 function getBackstageLiveSyncTimeoutMs(): number {
@@ -2040,7 +1893,11 @@ export function createPackageRoutes(auth: Auth, config: PackagesConfig) {
         },
         exp: Math.floor(Date.now() / 1000) + 3600,
       }) satisfies BackstageMaterializeClaims;
-      const materializeResult = await runBackstageMaterialize(config, materializeClaims);
+      const materializeResult = await runBackstageMaterialize({
+        ingestBaseUrl: config.ingestBaseUrl,
+        ingestSecret: config.backstageIngestSecret,
+        claims: materializeClaims,
+      });
       if (
         materializeResult.authUserId !== viewer.authUserId ||
         materializeResult.packageId !== packageId ||
